implement socket_vmnet
