+++ pkg/minikube/driver/driver.go
@@ -197,9 +197,6 @@ func AllowsPreload(driverName string) bool {
 
 // NeedsPortForward returns true if driver is unable provide direct IP connectivity
 func NeedsPortForward(name string) bool {
-	if IsQEMU(name) {
-		return true
-	}
 	if !IsKIC(name) {
 		return false
 	}
