+++ pkg/minikube/node/start.go
@@ -64,6 +64,7 @@ import (
 	"k8s.io/minikube/pkg/minikube/registry"
 	"k8s.io/minikube/pkg/minikube/style"
 	"k8s.io/minikube/pkg/minikube/vmpath"
+	"k8s.io/minikube/pkg/network"
 	"k8s.io/minikube/pkg/util"
 	"k8s.io/minikube/pkg/util/retry"
 	kconst "k8s.io/minikube/third_party/kubeadm/app/constants"
@@ -561,7 +562,7 @@ func startMachine(cfg *config.ClusterConfig, node *config.Node, delOnFail bool) 
 		return runner, preExists, m, host, errors.Wrap(err, "Failed to validate network")
 	}
 
-	if driver.IsQEMU(host.Driver.DriverName()) {
+	if driver.IsQEMU(host.Driver.DriverName()) && network.IsUser(cfg.Network) {
 		apiServerPort, err := getPort()
 		if err != nil {
 			return runner, preExists, m, host, errors.Wrap(err, "Failed to find apiserver port")
@@ -664,7 +665,7 @@ func validateNetwork(h *host.Host, r command.Runner, imageRepository string, kub
 		}
 	}
 
-	if !driver.BareMetal(h.Driver.DriverName()) && !driver.IsKIC(h.Driver.DriverName()) && !driver.IsQEMU(h.Driver.DriverName()) {
+	if shouldTrySSH(h.Driver.DriverName(), ip) {
 		if err := trySSH(h, ip); err != nil {
 			return ip, err
 		}
@@ -673,6 +674,17 @@ func validateNetwork(h *host.Host, r command.Runner, imageRepository string, kub
 	// Non-blocking
 	go tryRegistry(r, h.Driver.DriverName(), imageRepository, kubernetesVersion)
 	return ip, nil
+}
+
+func shouldTrySSH(driverName, ip string) bool {
+	if driver.BareMetal(driverName) || driver.IsKIC(driverName) {
+		return false
+	}
+	// QEMU with user network
+	if driver.IsQEMU(driverName) && ip == "127.0.0.1" {
+		return false
+	}
+	return true
 }
 
 func trySSH(h *host.Host, ip string) error {
