+++ pkg/minikube/registry/drvs/qemu2/qemu2.go
@@ -17,6 +17,7 @@ limitations under the License.
 package qemu2
 
 import (
+	"crypto/rand"
 	"fmt"
 	"os"
 	"os/exec"
@@ -158,10 +159,9 @@ func configure(cc config.ClusterConfig, n config.Node) (interface{}, error) {
 	if err != nil {
 		return nil, err
 	}
-	qemuNetwork := cc.Network
-	if qemuNetwork == "" {
-		qemuNetwork = "user"
-		// TODO: on next minor release, default to "socket".
+	mac, err := generateMACAddress()
+	if err != nil {
+		return nil, fmt.Errorf("generating MAC address: %v", err)
 	}
 
 	return qemu.Driver{
@@ -183,9 +183,10 @@ func configure(cc config.ClusterConfig, n config.Node) (interface{}, error) {
 		CPUType:        qemuCPU,
 		Firmware:       qemuFirmware,
 		VirtioDrives:   false,
-		Network:        qemuNetwork,
+		Network:        cc.Network,
 		CacheMode:      "default",
 		IOMode:         "threads",
+		MACAddress:     mac,
 	}, nil
 }
 
@@ -209,4 +210,15 @@ func status() registry.State {
 	}
 
 	return registry.State{Installed: true, Healthy: true, Running: true}
+}
+
+func generateMACAddress() (string, error) {
+	buf := make([]byte, 6)
+	if _, err := rand.Read(buf); err != nil {
+		return "", err
+	}
+	// Set local bit, ensure unicast address, socket_vmnet doesn't support multicast
+	buf[0] = (buf[0] | 2) & 0xfe
+	mac := fmt.Sprintf("%02x:%02x:%02x:%02x:%02x:%02x", buf[0], buf[1], buf[2], buf[3], buf[4], buf[5])
+	return mac, nil
 }
