+++ translations/ja.json
@@ -641,6 +641,7 @@
 	"The control plane node is not running (state={{.state}})": "コントロールプレーンノードは実行中ではありません (state={{.state}})",
 	"The control plane node must be running for this command": "このコマンドではコントロールプレーンノードが実行中でなければなりません",
 	"The cri socket path to be used.": "使用される CRI ソケットパス。",
+	"The default network for QEMU will change from 'user' to 'socket' in a future release": "",
 	"The docker-env command is incompatible with multi-node clusters. Use the 'registry' add-on: https://minikube.sigs.k8s.io/docs/handbook/registry/": "docker-env コマンドはマルチノードクラスターと互換性がありません。'registry' アドオンを使用してください: https://minikube.sigs.k8s.io/docs/handbook/registry/",
 	"The docker-env command is only compatible with the \"docker\" runtime, but this cluster was configured to use the \"{{.runtime}}\" runtime.": "docker-env コマンドは「docker」ランタイムとだけ互換性がありますが、このクラスターは「{{.runtime}}」ランタイムを使用するよう設定されています。",
 	"The driver '{{.driver}}' is not supported on {{.os}}/{{.arch}}": "'{{.driver}}' ドライバーは {{.os}}/{{.arch}} に対応していません",
@@ -905,9 +906,11 @@
 	"minikube provisions and manages local Kubernetes clusters optimized for development workflows.": "minikube は、開発ワークフロー用に最適化されたローカル Kubernetes クラスターを構築・管理します。",
 	"minikube quickly sets up a local Kubernetes cluster": "minikube はローカル Kubernetes クラスターを迅速にセットアップします",
 	"minikube service is not currently implemented with the qemu2 driver. See https://github.com/kubernetes/minikube/issues/14146 for details.": "minikube サービスは現在、qemu2 ドライバーでは実装されていません。詳細については、https://github.com/kubernetes/minikube/issues/14146 を参照してください。",
+	"minikube service is not currently implemented with the user network on QEMU, try starting minikube with '--network=socket'": "",
 	"minikube skips various validations when --force is supplied; this may lead to unexpected behavior": "minikube は --force が付与された場合、様々な検証をスキップします (これは予期せぬ挙動を引き起こすかも知れません)",
 	"minikube status --output OUTPUT. json, text": "minikube status --output OUTPUT. json, text",
 	"minikube tunnel is not currently implemented with the qemu2 driver. See https://github.com/kubernetes/minikube/issues/14146 for details.": "minikube トンネルは現在、qemu2 ドライバーでは実装されていません。 詳細については、https://github.com/kubernetes/minikube/issues/14146 を参照してください。",
+	"minikube tunnel is not currently implemented with the user network on QEMU, try starting minikube with '--network=socket'": "",
 	"minikube {{.version}} is available! Download it: {{.url}}": "minikube {{.version}} が利用可能です！次の URL からダウンロードしてください: {{.url}}",
 	"mkcmp is used to compare performance of two minikube binaries": "mkcmp で 2 つの minikube のバイナリーのパフォーマンスを比較できます",
 	"mount argument \"{{.value}}\" must be in form: \u003csource directory\u003e:\u003ctarget directory\u003e": "マウント引数「{{.value}}」は次の形式でなければなりません: \u003cソースディレクトリー\u003e:\u003cターゲットディレクトリー\u003e",
@@ -930,7 +933,6 @@
 	"scheduled stop is not supported on the none driver, skipping scheduling": "none ドライバーでは予定停止がサポートされていません (予約をスキップします)",
 	"service {{.namespace_name}}/{{.service_name}} has no node port": "サービス {{.namespace_name}}/{{.service_name}} は NodePort がありません",
 	"set tunnel bind address, empty or '*' indicates the tunnel should be available for all interfaces": "トンネル バインド アドレスを設定します。空または '*' は、トンネルがすべてのインターフェイスで使用可能であることを示します",
-	"socket_vmnet network flags are not yet implemented with the qemu2 driver.\n    See https://github.com/kubernetes/minikube/pull/14890 for details.": "",
 	"stat failed": "stat に失敗しました",
 	"status json failure": "status json に失敗しました",
 	"status text failure": "status text に失敗しました",
