+++ translations/ru.json
@@ -637,6 +637,7 @@
 	"The control plane node is not running (state={{.state}})": "",
 	"The control plane node must be running for this command": "",
 	"The cri socket path to be used.": "",
+	"The default network for QEMU will change from 'user' to 'socket' in a future release": "",
 	"The docker-env command is incompatible with multi-node clusters. Use the 'registry' add-on: https://minikube.sigs.k8s.io/docs/handbook/registry/": "",
 	"The docker-env command is only compatible with the \"docker\" runtime, but this cluster was configured to use the \"{{.runtime}}\" runtime.": "",
 	"The driver '{{.driver}}' is not supported on {{.os}}/{{.arch}}": "",
@@ -900,10 +901,10 @@
 	"minikube profile was successfully set to {{.profile_name}}": "",
 	"minikube provisions and manages local Kubernetes clusters optimized for development workflows.": "",
 	"minikube quickly sets up a local Kubernetes cluster": "",
-	"minikube service is not currently implemented with the qemu2 driver. See https://github.com/kubernetes/minikube/issues/14146 for details.": "",
+	"minikube service is not currently implemented with the user network on QEMU, try starting minikube with '--network=socket'": "",
 	"minikube skips various validations when --force is supplied; this may lead to unexpected behavior": "",
 	"minikube status --output OUTPUT. json, text": "",
-	"minikube tunnel is not currently implemented with the qemu2 driver. See https://github.com/kubernetes/minikube/issues/14146 for details.": "",
+	"minikube tunnel is not currently implemented with the user network on QEMU, try starting minikube with '--network=socket'": "",
 	"minikube {{.version}} is available! Download it: {{.url}}": "",
 	"mkcmp is used to compare performance of two minikube binaries": "",
 	"mount argument \"{{.value}}\" must be in form: \u003csource directory\u003e:\u003ctarget directory\u003e": "",
@@ -926,7 +927,6 @@
 	"scheduled stop is not supported on the none driver, skipping scheduling": "",
 	"service {{.namespace_name}}/{{.service_name}} has no node port": "",
 	"set tunnel bind address, empty or '*' indicates the tunnel should be available for all interfaces": "",
-	"socket_vmnet network flags are not yet implemented with the qemu2 driver.\n    See https://github.com/kubernetes/minikube/pull/14890 for details.": "",
 	"stat failed": "",
 	"status json failure": "",
 	"status text failure": "",
